Add progress checking to uploads
an endpoint to check if a chunk has been uploaded. not sure if this is necessary. will prob use localstorage to track it to reduce no. requests made

new state to track files queued for upload
- name
- progress
- state

still need to add resumable uploads and the option to cancel an upload and also delete uploaded file